Backend listo! espero que funcione porque probé muy poco.

+++ app/CollaborativeRecommenderSystem/Rating.java
@@ -19,7 +19,7 @@ public class Rating {
     }
 
     private Movie movie;
-    public Rating(int movieP,double newRating)
+    public Rating(long movieP,double newRating)
     {
         movie=Movie.find(movieP);
         rating=newRating;
